face-color-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
Pillow
openai
python-dotenv
serpapi
opencv-python
mediapipe
scikit-learn
numpy 